refactor: move split views
