+++ src/components/layout/split-views/side-by-side-split-view/rr-side-by-side-split-view.ts
@@ -15,7 +15,7 @@
 
 import { LitElement, html, css } from 'lit';
 import { customElement } from 'lit/decorators.js';
-import './rr-split-view-divider.ts';
+import '../split-view-divider/rr-split-view-divider.ts';
 
 @customElement('rr-side-by-side-split-view')
 export class RRSideBySideSplitView extends LitElement {
