+++ src/components/layout/split-views/split-view-pane/rr-split-view-pane.ts
