+++ src/components/layout/split-views/stacked-split-view/rr-stacked-split-view.ts
@@ -15,7 +15,7 @@
 
 import { LitElement, html, css } from 'lit';
 import { customElement } from 'lit/decorators.js';
-import './rr-split-view-divider.ts';
+import '../split-view-divider/rr-split-view-divider.ts';
 
 @customElement('rr-stacked-split-view')
 export class RRStackedSplitView extends LitElement {
